+++ src/database/migrations/2019_03_16_005634_create_groups_table.php
@@ -18,7 +18,7 @@ class CreateGroupsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name')->unique()->nullable(false);
             $table->string('slug')->unique()->nullable(false);
-            $table->text('description')->nullable(false);
+            $table->text('description')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });